requirements.txt
discord.py[voice]
wavelink==2.3.1